Implement steady fps.
The function steady_fps receive the desired fps and wait the required
wait time between calls.

For some reason, yet to be investigated, the fps counting seems to be
doubled. When trying to keep steady at 30 fps, it was showing 60 fps,
so for keeping the 30 fps the function was called for 15 fps.